🎨 Add apollo service file
